quz-01 - Extracting WidgetQuestions component
